pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketsense"
version = "0.1.0"
description = "Plataforma de analista financiero impulsada por IA usando agentes LangGraph en AWS Bedrock"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Giancarlos" }]

dependencies = [
    # ── Orquestación de LLMs ──────────────────────────────────
    "langgraph>=1.1,<2",
    "langchain>=1.2,<2",
    "langchain-google-genai",       # Integración LangChain <-> Google Gemini
    "langchain-community",          # Cargadores de documentos, vector stores, etc.

    # ── SDK de AWS ────────────────────────────────────────────
    "boto3",                        # Bedrock Runtime, RDS, S3

    # ── Base de Datos ─────────────────────────────────────────
    "sqlalchemy>=2.0,<3",
    "psycopg2-binary",              # Driver de PostgreSQL

    # ── API ────────────────────────────────────────────────────
    "fastapi",
    "uvicorn[standard]",            # Servidor ASGI

    # ── Datos y Configuración ─────────────────────────────────
    "pydantic>=2.0,<3",
    "python-dotenv",

    # ── RAG / Embeddings ──────────────────────────────────────
    "faiss-cpu",                    # Vector store local para desarrollo

    # ── Datos y ETL ───────────────────────────────────────────
    "pandas",
    "yfinance",                     # Datos de mercado del S&P 500
    "requests",                     # Llamadas HTTP (APIs de noticias, etc.)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff>=0.9",
    "jupyter",
    "ipykernel",
    "httpx",                        # Cliente de pruebas asíncrono para FastAPI
]

[tool.setuptools.packages.find]
include = ["backend*", "tests*"]

# ── Ruff (linter + formateador) ───────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",    # errores de pycodestyle
    "W",    # advertencias de pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── Pytest ─────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
